refactor: move login and register to auth API
