+++ VBWEB/backend/src/app.module.ts
@@ -3,14 +3,16 @@ import { Module } from '@nestjs/common';
 import { UserModule } from './user/user.module';
 import { SupabaseModule } from './supabase/supabase.module';
 import { ConfigModule } from '@nestjs/config';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
-    UserModule,
     SupabaseModule,
+    AuthModule,
+    UserModule,
   ]
 })
 export class AppModule {}
